Overlays_Test
Playing with overlay elements, can imagine an X-Axis locked GSAP Camera animation up button click.
Might be worth considering a lander like this given some r&d I did on the other design.

I'm a little concerned about the mobile experience with this approach tho. And losing the plot of "connect people around this project," but we will design it toward that.